IEA Upstream emissions estimates oil and gas
Takes the IEA data set released in January 2021 and arranges them in order of carbon intensity to give percentiles - P5, weighted average, and P95 estimates of CO2E for oil and gas. Note that for oil Opgee is used since it is country specific but IEA global oil emissions intensities included for reference.